Наладил архитектуру
