+++ app/src/main/java/com/gb/stopwatch/data/di/modules.kt
@@ -1,0 +1,9 @@
+package com.gb.stopwatch.data.di
+
+import com.gb.stopwatch.presentation.viewModel.MainActivityViewModel
+import com.gb.stopwatch.presentation.viewModel.MainActivityViewModelContract
+import org.koin.dsl.module
+
+val mainModule = module {
+    single<MainActivityViewModelContract.ViewModel> { MainActivityViewModel() }
+}
