+++ app/src/main/java/com/gb/stopwatch/domain/state/StopwatchState.kt
@@ -1,0 +1,13 @@
+package com.gb.stopwatch.domain.state
+
+sealed class StopwatchState {
+
+    data class Paused(
+        val elapsedTime: Long
+    ) : StopwatchState()
+
+    data class Running(
+        val startTime: Long,
+        val elapsedTime: Long
+    ) : StopwatchState()
+}
