+++ app/src/main/java/com/gb/stopwatch/domain/state/StopwatchStateHolder.kt
@@ -1,0 +1,34 @@
+package com.gb.stopwatch.domain.state
+
+import com.gb.stopwatch.domain.utils.ElapsedTimeCalculator
+import com.gb.stopwatch.domain.utils.StopwatchStateCalculator
+import com.gb.stopwatch.domain.utils.TimestampMillisecondsFormatter
+
+class StopwatchStateHolder(
+    private val stopwatchStateCalculator: StopwatchStateCalculator,
+    private val elapsedTimeCalculator: ElapsedTimeCalculator,
+    private val timestampMillisecondsFormatter: TimestampMillisecondsFormatter
+) {
+
+    private var currentState: StopwatchState = StopwatchState.Paused(0)
+
+    fun start() {
+        currentState = stopwatchStateCalculator.calculateRunningState(currentState)
+    }
+
+    fun pause() {
+        currentState = stopwatchStateCalculator.calculatePausedState(currentState)
+    }
+
+    fun stop() {
+        currentState = StopwatchState.Paused(0)
+    }
+
+    fun getStringTimeRepresentation(): String {
+        val elapsedTime = when (val currentState = currentState) {
+            is StopwatchState.Paused -> currentState.elapsedTime
+            is StopwatchState.Running -> elapsedTimeCalculator.calculate(currentState)
+        }
+        return timestampMillisecondsFormatter.format(elapsedTime)
+    }
+}
