+++ app/src/main/java/com/gb/stopwatch/domain/utils/ElapsedTimeCalculator.kt
@@ -1,0 +1,19 @@
+package com.gb.stopwatch.domain.utils
+
+import com.gb.stopwatch.domain.state.StopwatchState
+
+
+class ElapsedTimeCalculator(
+    private val timestampProvider: TimestampProvider,
+) {
+
+    fun calculate(state: StopwatchState.Running): Long {
+        val currentTimestamp = timestampProvider.getMilliseconds()
+        val timePassedSinceStart = if (currentTimestamp > state.startTime) {
+            currentTimestamp - state.startTime
+        } else {
+            0
+        }
+        return timePassedSinceStart + state.elapsedTime
+    }
+}
