+++ app/src/main/java/com/gb/stopwatch/domain/utils/TimestampProvider.kt
@@ -1,0 +1,5 @@
+package com.gb.stopwatch.domain.utils
+
+interface TimestampProvider {
+    fun getMilliseconds(): Long
+}
