+++ app/src/main/java/com/gb/stopwatch/presentation/viewModel/MainActivityViewModelContract.kt
@@ -1,0 +1,16 @@
+package com.gb.stopwatch.presentation.viewModel
+
+import kotlinx.coroutines.flow.StateFlow
+
+interface MainActivityViewModelContract {
+    abstract class ViewModel : androidx.lifecycle.ViewModel() {
+        abstract fun onStopwatchStartClick()
+
+        abstract fun onStopwatchStartPause()
+
+        abstract fun onStopwatchStartStop()
+
+        abstract val stopwatchState: StateFlow<String>
+
+    }
+}
